db: update ATAR distribution schema to float64 and mount 2025 data
- Modify `statistical_value` column in `atar_distributions` table from `DECIMAL(20, 15)` to `DOUBLE PRECISION` (`Float` in SQLAlchemy) to support full IEEE-754 float64 precision.
- Mount the new 2025 SQL initialization files in both `docker-compose.yml` and `docker-compose.dev.yml` for automated setup.

+++ backend/app/models/standard_models.py
@@ -39,9 +39,10 @@ class StandardWeighting(Base):
 
 class ATARDistribution(Base):
     __tablename__ = "atar_distributions"
-    academic_year = Column(Integer, primary_key=True)
-    statistical_value = Column(DECIMAL(20, 15), primary_key=True)
-    frequency = Column(Integer, nullable=False)
+    academic_year     = Column(Integer, primary_key=True)
+    statistical_value = Column(Float, primary_key=True)   # DOUBLE PRECISION - full float64 precision
+    frequency         = Column(Integer, nullable=False)
+
 
 
 class ParticipationRate(Base):
